Move `tests` to root
Change how resources are loaded and update `import`s

- - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #99
