+++ package.json
@@ -5,13 +5,13 @@
   "description": "Best practices tool for the web",
   "main": "./dist/lib/sonar.js",
   "scripts": {
-    "build:assets": "cpx \"./src/**/*.!(ts)\" dist",
-    "build:ts": "tsc --outDir dist --rootDir src",
+    "build:assets": "cpx \"./{src,tests}/**/*.!(ts)\" dist",
+    "build:ts": "tsc --outDir dist",
     "build": "npm run clean && npm-run-all build:*",
     "clean": "rimraf dist",
     "lint": "eslint --ext ts src/ --ext md . --ignore-pattern dist/*",
     "site": "node dist/bin/sonar --debug",
-    "test": "npm run lint && npm run clean && npm run build && nyc ava",
+    "test": "npm run lint && npm run build && nyc ava",
     "watch": "npm run build && npm-run-all --parallel -c watch:*",
     "watch:ts": "npm run build:ts -- --watch",
     "watch:test": "ava --watch",
