+++ src/lib/util/resource-loader.ts
@@ -30,7 +30,7 @@ const TYPE = {
 
 /** Loads all the resources available for the given type */
 const loadOfType = (type: string): Map<string, Resource> => {
-    let resourceFiles: string[] = globby.sync(`{./,./node_modules/sonar-*}dist/lib/${type}s/**/*.js`,
+    let resourceFiles: string[] = globby.sync(`{./,./node_modules/sonar-*}dist/src/lib/${type}s/**/*.js`,
         { absolute: true });
 
     resourceFiles = resourceFiles.filter((resourceFile) => {
