--- src/tests/lib/collectors/cdp.ts
@@ -1,7 +1,0 @@
-import builder from '../../../lib/collectors/cdp/cdp';
-import { testCollector } from './_common';
-
-// First test that it passes the standard tests for all collectors
-testCollector(builder);
-
-// Specific tests go here
