--- src/tests/lib/collectors/jsdom.ts
@@ -1,7 +1,0 @@
-import builder from '../../../lib/collectors/jsdom/jsdom';
-import { testCollector } from './_common';
-
-// First test that it passes the standard tests for all collectors
-testCollector(builder);
-
-// Specific tests go here
