+++ tests/lib/collectors/fixtures/file-protocol.txt
