+++ tests/lib/collectors/jsdom.ts
@@ -1,0 +1,7 @@
+import builder from '../../../src/lib/collectors/jsdom/jsdom';
+import { testCollector } from './_common';
+
+// First test that it passes the standard tests for all collectors
+testCollector(builder);
+
+// Specific tests go here
