+++ tests/lib/util/file-loader/fixtures/fixture.js
