+++ tests/lib/util/fixtures/bom.txt
