+++ tsconfig.json
@@ -11,6 +11,7 @@
         "node_modules"
     ],
     "include": [
-        "src/**/*.ts"
+        "src/**/*.ts",
+        "tests/**/*.ts"
     ]
 }
